chore: organize ci unit tests
