+++ packages/ci/test/unit/commands/ci/info.unit.test.ts
