+++ packages/ci/test/unit/commands/ci/run.unit.test.ts
@@ -1,7 +1,7 @@
 import {expect, test} from '@oclif/test'
 import * as fs from 'async-file'
 
-import * as git from '../../../src/utils/git'
+import * as git from '../../../../src/utils/git'
 const got = require('got')
 
 describe('ci:run', () => {
